server/requirements.txt
django>=4.2,<5.2
djangorestframework>=3.15
django-cors-headers>=4.3
channels[daphne]>=4.0
PyJWT>=2.8
requests>=2.31
whitenoise>=6.7
